rename `frum` to `vector`
