+++ src/views/QuantumTP6/index.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { withStyles } from '@material-ui/core';
 import PropTypes from 'prop-types';
 import Grid from '@material-ui/core/Grid';
-import { frum } from '../../vendor/queryOps';
+import { vector } from '../../vendor/vectors';
 import { TP6_TESTS, TP6_PAGES } from '../../quantum/config';
 import { withNavigation } from '../../vendor/utils/navigation';
 import Picker from '../../vendor/utils/navigation/Picker';
@@ -25,7 +25,7 @@ class TP6 extends React.Component {
   render() {
     const { classes, navigation, test, bits } = this.props;
     const subtitle = `${
-      frum(TP6_TESTS)
+      vector(TP6_TESTS)
         .where({ id: test })
         .first().label
     } on ${bits} bits`;
@@ -37,7 +37,7 @@ class TP6 extends React.Component {
         <DashboardPage key={subtitle} title="TP6 Desktop" subtitle={subtitle}>
           {navigation}
           <Grid container spacing={24}>
-            {frum(TP6_PAGES)
+            {vector(TP6_PAGES)
               .where({ bits })
               .groupBy('title')
               .map((series, title) => (
@@ -48,7 +48,7 @@ class TP6 extends React.Component {
                   className={classes.chart}>
                   <PerfherderGraphContainer
                     title={title}
-                    series={frum(series)
+                    series={vector(series)
                       .sortBy(['browser'])
                       .reverse()
                       .map(s => ({
